mconfigptr: remove note about emulation
The same note could be added to all new CSRs.  Other CRSs don't mention
it because implementations are already given maximal freedom.
How the CSR is implemented is irrelevant to M-mode software.

The best UD can do is to leave the implementation details unspecified.
Implementers will know how to deal with it.

Signed-off-by: Radim Krčmář <rkrcmar@ventanamicro.com>